+++ packages/wallets-helpers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/wallets-helpers
 
+## 2.1.0
+
+### Minor Changes
+
+- Support EIP-6963
+
 ## 2.0.2
 
 ### Patch Changes